backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP client for OpenRouter
httpx>=0.25.0

# Microsoft Presidio for PII detection
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# GLiNER for transformer-based PII detection
gliner>=0.2.0
torch>=2.0.0

# SpaCy NLP model (small for Vercel 250MB limit)
spacy>=3.7.0

# Environment variables
python-dotenv>=1.0.0

# CORS middleware
starlette>=0.27.0
